+++ source/utils/variant.h
@@ -15,14 +15,14 @@ typedef struct
     VarianType type;
 
     union {
-        int as_int;
+        long int as_int;
         float as_float;
     };
 
     char as_string[VARIANT_STRING_SIZE];
 } Variant;
 
-Variant vint(int value);
+Variant vint(long int value);
 
 Variant vfloat(float value);
